Histogram/requirements.txt
Flask==1.0.2
matplotlib==3.2.1
requests==2.23.0
google-cloud-storage